refactor(st-ddr): reduce DDR driver footprint for STM32MP2
Remove useless code and files.
Reduce global variables notably by using them as entry parameters for
sub-functions.
Replace some very rarely used enum types by definitions.
Consider only single pstate case in PHY driver.
Fix MISRA and coding rules warnings.
Simplify some algorithms.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/365370
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>

+++ drivers/st/ddr/phy/phyinit/src/ddrphy_phyinit_writeoutmem.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021-2023, STMicroelectronics - All Rights Reserved
+ * Copyright (C) 2021-2024, STMicroelectronics - All Rights Reserved
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -25,22 +25,23 @@
  * @param[in] mem_size size of the memroy (in mem array index)
  * @returns void
  */
-void ddrphy_phyinit_writeoutmem(uint32_t *mem, int mem_offset, int mem_size)
+void ddrphy_phyinit_writeoutmem(uint32_t *mem, uint32_t mem_offset, uint32_t mem_size)
 {
-	int index;
+	uint32_t index;
 
 	/*
 	 * 1. Enable access to the internal CSRs by setting the MicroContMuxSel CSR to 0.
 	 *    This allows the memory controller unrestricted access to the configuration CSRs.
 	 */
-	mmio_write_16((uintptr_t)(DDRPHYC_BASE + 4 * (TAPBONLY | CSR_MICROCONTMUXSEL_ADDR)), 0x0U);
+	mmio_write_16((uintptr_t)(DDRPHYC_BASE + (4U * (TAPBONLY | CSR_MICROCONTMUXSEL_ADDR))),
+		      0x0U);
 
-	for (index = 0; index < (int)(mem_size / sizeof(uint32_t)); index++) {
+	for (index = 0U; index < mem_size / sizeof(uint32_t); index++) {
 		uint32_t data = mem[index];
 
-		mmio_write_16((uintptr_t)(DDRPHYC_BASE + 4 * ((index * 2) + mem_offset)),
+		mmio_write_16((uintptr_t)(DDRPHYC_BASE + (4U * ((index * 2) + mem_offset))),
 			     data & 0xFFFFU);
-		mmio_write_16((uintptr_t)(DDRPHYC_BASE + 4 * ((index * 2) + 1 + mem_offset)),
+		mmio_write_16((uintptr_t)(DDRPHYC_BASE + (4U * ((index * 2) + 1 + mem_offset))),
 			     (data >> 16) & 0xFFFFU);
 	}
 
@@ -48,30 +49,31 @@ void ddrphy_phyinit_writeoutmem(uint32_t *mem, int mem_offset, int mem_size)
 	 * 2. Isolate the APB access from the internal CSRs by setting the MicroContMuxSel CSR to 1.
 	 *    This allows the firmware unrestricted access to the configuration CSRs.
 	 */
-	mmio_write_16((uintptr_t)(DDRPHYC_BASE + 4 * (TAPBONLY | CSR_MICROCONTMUXSEL_ADDR)), 0x1U);
+	mmio_write_16((uintptr_t)(DDRPHYC_BASE + (4U * (TAPBONLY | CSR_MICROCONTMUXSEL_ADDR))),
+		      0x1U);
 }
 
-/*
- * Similar function for message block
- */
-void ddrphy_phyinit_writeoutmsgblk(uint16_t *mem, int mem_offset, int mem_size)
+/* Similar function for message block */
+void ddrphy_phyinit_writeoutmsgblk(uint16_t *mem, uint32_t mem_offset, uint32_t mem_size)
 {
-	int index;
+	uint32_t index;
 
 	/*
 	 * 1. Enable access to the internal CSRs by setting the MicroContMuxSel CSR to 0.
 	 *    This allows the memory controller unrestricted access to the configuration CSRs.
 	 */
-	mmio_write_16((uintptr_t)(DDRPHYC_BASE + 4 * (TAPBONLY | CSR_MICROCONTMUXSEL_ADDR)), 0x0U);
+	mmio_write_16((uintptr_t)(DDRPHYC_BASE + (4U * (TAPBONLY | CSR_MICROCONTMUXSEL_ADDR))),
+		      0x0U);
 
 
-	for (index = 0; index < (int)(mem_size / sizeof(uint16_t)); index++) {
-		mmio_write_16((uintptr_t)(DDRPHYC_BASE + 4 * (index + mem_offset)), mem[index]);
+	for (index = 0U; index < mem_size / sizeof(uint16_t); index++) {
+		mmio_write_16((uintptr_t)(DDRPHYC_BASE + (4U * (index + mem_offset))), mem[index]);
 	}
 
 	/*
 	 * 2. Isolate the APB access from the internal CSRs by setting the MicroContMuxSel CSR to 1.
 	 *    This allows the firmware unrestricted access to the configuration CSRs.
 	 */
-	mmio_write_16((uintptr_t)(DDRPHYC_BASE + 4 * (TAPBONLY | CSR_MICROCONTMUXSEL_ADDR)), 0x1U);
+	mmio_write_16((uintptr_t)(DDRPHYC_BASE + (4U * (TAPBONLY | CSR_MICROCONTMUXSEL_ADDR))),
+		      0x1U);
 }
